Add files via upload
I created a model which recognized that a given email is spam or not. The major tools used for this are Python, Machine Learning, Deep Learning. for model building I used more than 5 classifiers to see which one gives good accuracy. RandomForest, XGB SVC and MultinomialNB classfier . After using voting classifier I got the the accuracy with 0.98 and precision with 0.99. So, Multinomial NB gives the best accuracy with precision.